+++ contrib/docker/build.sh
@@ -27,6 +27,9 @@ if [[ -z "$1" ]]; then
   target=image
   outputFmt='-o type=docker,type=registry'
   tagsuffix=':devel'
+  if [[ -n "$GIT_BRANCH" ]] && [[ "x$GIT_BRANCH" != "xmaster" ]]; then
+    tagsuffix="$tagsuffix-$GIT_BRANCH"
+  fi
 elif [[ "x$1" = "xrelease" ]]; then
   namesuffix='.release'
   target=image
@@ -58,6 +61,7 @@ for image in $images; do
     --build-arg "UPLOAD_URL=$UPLOAD_URL" \
     --build-arg "UPLOAD_CREDENTIALS=$UPLOAD_CREDENTIALS" \
     --build-arg "UPLOAD_OS=$image" \
+    --build-arg "GIT_REVISION=$GIT_REVISION" \
     -t $tagprefix$tagsuffix \
     $extratag \
     $output \