+++ src/ebusd/mqtthandler.cpp
@@ -150,7 +150,7 @@ void splitFields(const string& str, vector<string>* row);
  */
 static error_t mqtt_parse_opt(int key, char *arg, struct argp_state *state) {
   result_t result = RESULT_OK;
-
+  unsigned int value;
   switch (key) {
   case O_HOST:  // --mqtthost=localhost
     if (arg == nullptr || arg[0] == 0) {
@@ -161,11 +161,12 @@ static error_t mqtt_parse_opt(int key, char *arg, struct argp_state *state) {
     break;
 
   case O_PORT:  // --mqttport=1883
-    g_port = (uint16_t)parseInt(arg, 10, 1, 65535, &result);
+    value = parseInt(arg, 10, 1, 65535, &result);
     if (result != RESULT_OK) {
       argp_error(state, "invalid mqttport");
       return EINVAL;
     }
+    g_port = (uint16_t)value;
     break;
 
   case O_CLID:  // --mqttclientid=clientid
@@ -193,28 +194,28 @@ static error_t mqtt_parse_opt(int key, char *arg, struct argp_state *state) {
     break;
 
   case O_TOPI:  // --mqtttopic=ebusd
+  {
     if (arg == nullptr || arg[0] == 0 || strchr(arg, '+') || arg[strlen(arg)-1] == '/') {
       argp_error(state, "invalid mqtttopic");
       return EINVAL;
-    } else {
-      char *pos = strchr(arg, '#');
-      if (pos && (pos == arg || pos[1])) {  // allow # only at very last position (to indicate not using any default)
-        argp_error(state, "invalid mqtttopic");
-        return EINVAL;
-      }
+    }
+    char *pos = strchr(arg, '#');
+    if (pos && (pos == arg || pos[1])) {  // allow # only at very last position (to indicate not using any default)
+      argp_error(state, "invalid mqtttopic");
+      return EINVAL;
     }
     if (g_topic) {
       argp_error(state, "duplicate mqtttopic");
       return EINVAL;
-    } else {
-      StringReplacer replacer;
-      if (!replacer.parse(arg, true)) {
-        argp_error(state, "malformed mqtttopic");
-        return EINVAL;
-      }
-      g_topic = arg;
     }
+    StringReplacer replacer;
+    if (!replacer.parse(arg, true)) {
+      argp_error(state, "malformed mqtttopic");
+      return ESRCH;  // abort in any case due to the above potentially being destructive
+    }
+    g_topic = arg;
     break;
+  }
 
   case O_GTOP:  // --mqttglobal=global/
     if (arg == nullptr || strchr(arg, '+') || strchr(arg, '#')) {
@@ -229,11 +230,12 @@ static error_t mqtt_parse_opt(int key, char *arg, struct argp_state *state) {
     break;
 
   case O_PQOS:  // --mqttqos=0
-    g_qos = parseSignedInt(arg, 10, 0, 2, &result);
+    value = parseInt(arg, 10, 0, 2, &result);
     if (result != RESULT_OK) {
       argp_error(state, "invalid mqttqos value");
       return EINVAL;
     }
+    g_qos = static_cast<signed>(value);
     break;
 
   case O_INTF:  // --mqttint=/etc/ebusd/mqttint.cfg
@@ -374,7 +376,8 @@ bool mqtthandler_register(UserInfo* userInfo, BusHandler* busHandler, MessageMap
     int revision = -1;
     mosquitto_lib_version(&major, &minor, &revision);
     if (major < LIBMOSQUITTO_MAJOR) {
-      logOtherError("mqtt", "invalid mosquitto version %d instead of %d, will try connecting anyway", major, LIBMOSQUITTO_MAJOR);
+      logOtherError("mqtt", "invalid mosquitto version %d instead of %d, will try connecting anyway", major,
+        LIBMOSQUITTO_MAJOR);
     }
     logOtherInfo("mqtt", "mosquitto version %d.%d.%d (compiled with %d.%d.%d)", major, minor, revision,
       LIBMOSQUITTO_MAJOR, LIBMOSQUITTO_MINOR, LIBMOSQUITTO_REVISION);
@@ -482,7 +485,8 @@ string removeTrailingNonTopicPart(const string& str) {
 
 MqttHandler::MqttHandler(UserInfo* userInfo, BusHandler* busHandler, MessageMap* messages)
   : DataSink(userInfo, "mqtt"), DataSource(busHandler), WaitThread(), m_messages(messages), m_connected(false),
-    m_initialConnectFailed(false), m_lastUpdateCheckResult("."), m_lastScanStatus(SCAN_STATUS_NONE), m_lastErrorLogTime(0) {
+    m_initialConnectFailed(false), m_lastUpdateCheckResult("."), m_lastScanStatus(SCAN_STATUS_NONE),
+    m_lastErrorLogTime(0) {
   m_definitionsSince = 0;
   m_mosquitto = nullptr;
   bool hasIntegration = false;
@@ -540,7 +544,7 @@ MqttHandler::MqttHandler(UserInfo* userInfo, BusHandler* busHandler, MessageMap*
     m_replacers.set("version", PACKAGE_VERSION);
     if (m_replacers["prefix"].empty()) {
       string line = m_replacers.get("topic", true);
-      if (line.empty() || line=="/") {
+      if (line.empty() || line == "/") {
         line = string(PACKAGE);  // ensure prefix if cmdline topic is absent
       }
       m_replacers.set("prefix", line);
@@ -1075,11 +1079,11 @@ void MqttHandler::run() {
             if (dataType->isNumeric()) {
               auto dt = dynamic_cast<const NumberDataType*>(dataType);
               ostr.str("");
-              if (dt->getMinMax(false, OF_NONE, &ostr) == RESULT_OK) {
+              if (dt->getMinMax(false, g_publishFormat, &ostr) == RESULT_OK) {
                 values.set("min", ostr.str());
                 ostr.str("");
               }
-              if (dt->getMinMax(true, OF_NONE, &ostr) == RESULT_OK) {
+              if (dt->getMinMax(true, g_publishFormat, &ostr) == RESULT_OK) {
                 values.set("max", ostr.str());
               }
             }