+++ src/lib/utils/httpclient.cpp
@@ -152,7 +152,7 @@ SSLSocket* SSLSocket::connect(const string& host, const uint16_t& port, bool htt
   time_t until = time(nullptr) + (timeout <= 3 ? 3 : timeout);  // at least 3 seconds
   if (!https) {
     do {
-      bio = BIO_new_connect(hostPort.c_str());
+      bio = BIO_new_connect((char*)hostPort.c_str());
       if (isError("connect", bio)) {
         break;
       }
@@ -170,7 +170,7 @@ SSLSocket* SSLSocket::connect(const string& host, const uint16_t& port, bool htt
       if (isError("ctx_new", ctx)) {
         break;
       }
-      bool verifyPeer = !caFile || strcmp(caFile, "#")!=0;
+      bool verifyPeer = !caFile || strcmp(caFile, "#") != 0;
       SSL_CTX_set_verify(ctx, verifyPeer ? SSL_VERIFY_PEER : SSL_VERIFY_NONE, nullptr);
       if (verifyPeer) {
 #if OPENSSL_VERSION_NUMBER >= 0x10101000L